Added profiling script and switched to f32 uv

+++ src/geo/mod.rs
@@ -6,12 +6,12 @@ pub mod vec3;
 
 #[derive(Copy, Clone, Debug)]
 pub struct Uv {
-    pub u: f64,
-    pub v: f64,
+    pub u: f32,
+    pub v: f32,
 }
 
 impl Uv {
-    pub fn new(u: f64, v: f64) -> Uv {
+    pub fn new(u: f32, v: f32) -> Uv {
         Uv { u, v }
     }
 }
